+++ Semana 19 - Arquitetura de Software/Projeto - LaBook/src/index.ts
@@ -9,6 +9,7 @@ import { authenticationData, user, post } from "./model/types";
 import { hashCreate } from "./services/hashManager";
 import { generateId } from "./services/generateId";
 import { generateToken } from "./services/authenticator";
+import { signup } from "./controller/user/signup";
 
 /**************************** CONFIG ******************************/
 
@@ -34,37 +35,7 @@ app.use(cors());
 
 /**************************** ENDPOINTS ******************************/
 
-app.post("/users/signup", async (req: Request, res: Response) => {
-  try {
-    let message = "Success!";
-    const { name, email, password } = req.body;
-
-    if (!name || !email || !password) {
-      res.statusCode = 406;
-      throw new Error('"name", "email" and "password" must be provided');
-    }
-
-    const id: string = generateId();
-
-    const hashPassword: string = hashCreate(password);
-
-    await connection.raw(`
-      INSERT INTO user_labook (id, name, email, password)
-      VALUES(
-        "${id}",
-        "${name}",
-        "${email}",
-        "${hashPassword}"
-      )
-    `)
-
-    const token: string = generateToken({ id })
-
-    res.status(201).send({ message, token })
-  } catch (error) {
-    res.status(400).send({ message: error.sqlMessage || error.message });
-  }
-});
+app.post("/users/signup", signup);
 
 // app.post("/users/login", async (req: Request, res: Response) => {
 //   try {
